Fix age input type and formatting in Exercicio6.py

+++ Exercicio6.py
@@ -1,11 +1,11 @@
 print("=== Indicador de Faixa Etaria ===")
 
-idade = float(input("Digite a Sua idade: "))
+idade = int(input("Digite a Sua idade: "))
 
 if idade < 18:
     print("Menor de idade !!!!")
 
-elif idade <60 >18:
+elif idade <60  >18:
     print("Adulto !!!!")
 
 else:
